requirements.txt
python-docx
pymongo